alert upon delete

+++ fvtracker/lib/utils/webstore/orders.js
@@ -26,6 +26,7 @@ export const fillOrdersRedux = async ({ dispatch }) => {
 };
 
 export const deleteOrder = async ({ orderId, dispatch, router }) => {
+  if (!alert("Jeste li sigurni da želite obrisati ovu narudžbu?")) return;
   try {
     await api.delete(`/orders`, {
       params: { id: orderId },
